requirements.txt
# 核心依赖
aiohttp>=3.9.0
aiohttp-socks>=0.8.0

# 数据库和数据处理
python-dotenv>=1.0.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Telegram功能 (可选)
pyTelegramBotAPI>=4.14.0
telethon>=1.34.0

# 图表和可视化 (可选，用于生成报告)
plotly>=5.18.0
dnspython==2.4.0
geoip2==4.7.0
tqdm==4.66.1
